requirements.txt
os
pandas
pymysql
sqlalchemy
python-dotenv
streamlit